fix workflows.yaml into directory.
