+++ testdata/workflows/retry-on-error.yaml
@@ -1,41 +1,3 @@
-# https://github.com/argoproj/argo/blob/master/examples/hello-world.yaml
-apiVersion: argoproj.io/v1alpha1
-kind: Workflow
-metadata:
-  creationTimestamp: "2020-03-14T12:22:56Z"
-  generateName: hello-world-
-  generation: 5
-  labels:
-    workflows.argoproj.io/completed: "true"
-    workflows.argoproj.io/phase: Succeeded
-  name: hello-world-f5s7c
-  namespace: default
-spec:
-  entrypoint: whalesay
-  templates:
-  - container:
-      args:
-      - hello world
-      command:
-      - cowsay
-      image: docker/whalesay:latest
-      name: ""
-    name: whalesay
-status:
-  finishedAt: "2020-03-14T12:23:14Z"
-  nodes:
-    hello-world-f5s7c:
-      displayName: hello-world-f5s7c
-      finishedAt: "2020-03-14T12:23:13Z"
-      id: hello-world-f5s7c
-      name: hello-world-f5s7c
-      phase: Succeeded
-      startedAt: "2020-03-14T12:22:56Z"
-      templateName: whalesay
-      type: Pod
-  phase: Succeeded
-  startedAt: "2020-03-14T12:22:56Z"
----
 # https://github.com/argoproj/argo/blob/master/examples/retry-on-error.yaml
 apiVersion: argoproj.io/v1alpha1
 kind: Workflow
